+++ analysis/quicklookXSlice.py
@@ -2,18 +2,28 @@ from MITgcmutils import mds
 from matplotlib import pyplot as plt
 import numpy as np
 import os
+import sys
 import cmocean
 import fileinput
 
-#Pick cross section
-crossSection = 1000 #default
-try:
-    with open('input/plotPoint.txt', 'r') as file:
-        lines = file.readlines()
-        crossSection = float(lines[3]) #reads the 3rd line in the doc
-        print('cross section read from file', crossSection)
-except FileNotFoundError:
-    print('plot point file does not exist, using default')
+# Pick cross section to view from file or default
+yCrossSection = 1000
+xCrossSection = 5000
+zDepth = -50
+plotDPI = 100
+cleanPNGs = True
+
+if(os.path.isfile('input/plotHelper.py')):
+    sys.path.append('input')
+    from plotHelper import *
+    print('Found experiment plotting settings')
+elif(os.path.isfile('../plotHelper.py')):
+    sys.path.append('../')
+    print('no custom plotting settings, using local default')
+    from plotHelper import *
+else:  
+    print('no defaults found')
+print('Plot DPI:',plotDPI,'; clean PNGs?',cleanPNGs)
 
 dt = 0.0   
 for line in fileinput.input('input/data'):
@@ -104,7 +114,7 @@ if(isBerg):
                     openFrac[:,j,i] = 1
 
 
-xSlice = np.argmin(np.abs(x[0,:] - crossSection))
+xSlice = np.argmin(np.abs(x[0,:] - xCrossSection))
 print('cross section is x =', x[0,xSlice],'index', xSlice)
 
 
@@ -179,11 +189,12 @@ for k in range(len(name)):
         
         str = "figs/sideX%s%05i.png" % (name[k],j)
         
-        plt.savefig(str, format='png')
+        plt.savefig(str, format='png', dpi=plotDPI)
         plt.close()
         #plt.show()
 
     os.system('magick -delay %f figs/sideX%s*.png -colors 256 -depth 256 figs/autosideX_%s.gif' %(500/((maxStep-startStep)/sizeStep), name[k], name[k]))
 
 #Clean up intermediate pngs
-os.system('rm -f figs/sideX*.png')
+if(cleanPNGs):
+    os.system('rm -f figs/sideX*.png')
